% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.R
\name{DiceRoll}
\alias{DiceRoll}
\title{Roll dice of your choosing}
\usage{
DiceRoll(m)
}
\arguments{
\item{m}{Dice of your choosing}
}
\value{
Dice Result
}
\description{
Roll a dice of your choosing (Description)
}
\details{
Roll a dice of your choosing and get a randomly generated number as the result (Details)
}
\author{
Dustin
}
