% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_dice.R
\name{parse_dice}
\alias{parse_dice}
\title{Parse a D&D style dice expression}
\usage{
parse_dice(expr)
}
\arguments{
\item{expr}{A string like "2d6+3" or "1d20"}
}
\value{
A list with components n, sides, modifier
}
\description{
Parse a D&D style dice expression
}
